requirements.txt
streamlit_option_menu
pandas~=1.5.2
numpy~=1.23.5
sklearn~=0.0.post1
scikit-learn~=1.2.2
imblearn~=0.0
streamlit~=1.36.0
Pillow~=9.3.0